Fix parsing of /network-profiles page
Sometime recently, http://zero.sjeng.org/network-profiles changed their
page, the links are no longer hard-coded in the HTML but instead
populated dynamically in Javascript. So the old regex was broken and
just never found a match.

Fix the regex to now grab the first hash from the inline JSON that Vue
uses to populate the table. Works on my own manual testing.