Fix saveBackRouting skipping nodes
`Board.saveBackRouting` is supposed to walk the list of `BoardHistoryNodes` from
the current position to the top of the tree, and at every node, set the
`fromBackChildren` of the previous `BoardHistoryNode` to the variation index
of the current node. However, due to a bug in `saveBackRouting`, half of the
nodes were being skipped in the process.

The effect of this is that when switching between different engines, the board
state could jump to a variation the user wasn't looking at previously. By
properly stepping through every node during `saveBackRouting`, switching between
engines always properly preserves the current variation.